requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
openai
python-multipart
python-dotenv
